requirements.txt
fastapi
uvicorn[standard]
psutil
python-dotenv
